Create database retrieval function
